remove PLUGIN_METABOT
