--- enterprise/frontend/src/metabase-enterprise/metabot/components/MetabotAdmin/MetabotPurchasePage/index.ts
@@ -1,1 +1,0 @@
-export * from "./MetabotPurchasePage";
