--- enterprise/frontend/src/metabase-enterprise/metabot/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useMetabotSQLSuggestion } from "./use-metabot-sql-suggestion";
