+++ enterprise/frontend/src/metabase-enterprise/plugins.ts
@@ -31,7 +31,6 @@ import { initializePlugin as initializeEmbeddingIframeSdk } from "./embedding_if
 import { initializePlugin as initializeEmbeddingIframeSdkSetup } from "./embedding_iframe_sdk_setup";
 import { initializePlugin as initializeFeatureLevelPermissions } from "./feature_level_permissions";
 import { initializePlugin as initializeGroupManagers } from "./group_managers";
-import { initializePlugin as initializeMetabot } from "./metabot";
 import { initializePlugin as initializeModelPersistence } from "./model_persistence";
 import { initializePlugin as initializeModeration } from "./moderation";
 import { initializePlugin as initializeRemoteSync } from "./remote_sync";
@@ -91,7 +90,6 @@ export function initializePlugins() {
   initializeResourceDownloads();
   initializeUserProvisioning();
   initializeCleanUp();
-  initializeMetabot();
   initializeDatabaseReplication();
   initializeTableEditing();
   initializeDependencies();
