+++ enterprise/frontend/src/metabase-enterprise/sdk-plugins.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/order */
 import { PLUGIN_IS_EE_BUILD } from "metabase/plugins";
 
 import { initializeHandleLinkPlugin } from "../embedding-sdk-ee/handle-link/handle-link";
@@ -11,13 +12,10 @@ import "./shared";
 import { initializePlugin as initializeContentTranslation } from "./content_translation";
 import { initializePlugin as initializeEmbedding } from "./embedding";
 import { initializePlugin as initializeEmbeddingSdk } from "./embedding-sdk";
-import { initializePlugin as initializeMetabot } from "./metabot";
 import { initializePlugin as initializeSharing } from "./sharing";
 import { initializePlugin as initializeTenants } from "./tenants";
 import { initializePlugin as initializeWhitelabelPlugin } from "./whitelabel";
 import { initializePlugin as initializeWhitelabelOverridePlugin } from "./whitelabel/sdk-overrides";
-
-// eslint-disable-next-line import/order -- This needs to be imported after Metabot plugin, otherwise it will initialize reducers before the Metabot plugin has set a reducer.
 import { initializePlugin as initializeNotifications } from "../embedding-sdk-ee/notifications";
 
 /**
@@ -27,7 +25,6 @@ import { initializePlugin as initializeNotifications } from "../embedding-sdk-ee
 export function initializePlugins() {
   initializeEmbedding?.();
   initializeEmbeddingSdk?.();
-  initializeMetabot?.();
   initializeTenants?.();
   initializeWhitelabelPlugin?.();
   initializeWhitelabelOverridePlugin?.();
