+++ enterprise/frontend/src/metabase-enterprise/urls.ts
@@ -24,10 +24,6 @@ export function removeDestinationDatabase(
   return `/admin/databases/${databaseId}/destination-databases/${destinationDatabaseId}/remove`;
 }
 
-export function newMetabotConversation({ prompt }: { prompt: string }) {
-  return `/metabot/new?q=${encodeURIComponent(prompt)}`;
-}
-
 export function newTenant() {
   return `/admin/people/tenants/new`;
 }
