+++ frontend/src/metabase/admin/settings/components/SettingsPages/SlackSettingsPage.tsx
@@ -1,7 +1,7 @@
 import { t } from "ttag";
 
 import { SettingsPageWrapper } from "metabase/admin/components/SettingsSection";
-import { PLUGIN_METABOT } from "metabase/plugins";
+import { MetabotSlackSetup } from "metabase/metabot/components/MetabotAdmin/MetabotSlackSetup";
 import { Icon } from "metabase/ui";
 
 import { SlackSetup } from "../../slack/SlackSetup";
@@ -17,7 +17,7 @@ export const SlackSettingsPage = () => {
         }
       >
         <SlackSetup />
-        <PLUGIN_METABOT.MetabotSlackSetup />
+        <MetabotSlackSetup />
       </SettingsPageWrapper>
     </>
   );
