+++ frontend/src/metabase/lib/urls/index.ts
@@ -10,6 +10,7 @@ export * from "./dependencies";
 export * from "./documents";
 export * from "./indexed-entities";
 export * from "./metrics";
+export * from "./metabot";
 export * from "./modelToUrl";
 export * from "./permissions";
 export * from "./misc";
