+++ frontend/src/metabase/lib/urls/metabot.ts
@@ -1,0 +1,3 @@
+export function newMetabotConversation({ prompt }: { prompt: string }) {
+  return `/metabot/new?q=${encodeURIComponent(prompt)}`;
+}
