+++ frontend/src/metabase/metabot/components/MetabotAdmin/MetabotAdminPage.tsx
@@ -1,29 +1,68 @@
-import { IndexRoute, Redirect } from "react-router";
-import { t } from "ttag";
+import { useDisclosure } from "@mantine/hooks";
+import type { ReactNode } from "react";
+import { IndexRoute, Redirect, Route } from "react-router";
+import { P, match } from "ts-pattern";
+import { c, t } from "ttag";
 
 import ErrorBoundary from "metabase/ErrorBoundary";
+import { SettingsSection } from "metabase/admin/components/SettingsSection";
+import { SettingHeader } from "metabase/admin/settings/components/SettingHeader";
 import {
-  AdminNavItem,
-  AdminNavWrapper,
-} from "metabase/admin/components/AdminNav";
+  skipToken,
+  useGetCollectionQuery,
+  useListMetabotsQuery,
+  useUpdateMetabotMutation,
+} from "metabase/api";
+import { useAdminSetting } from "metabase/api/utils/settings";
+import { canonicalCollectionId } from "metabase/collections/utils";
 import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
-import { PLUGIN_METABOT } from "metabase/plugins";
-import { Flex } from "metabase/ui";
+import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
+import { CollectionPickerModal } from "metabase/common/components/Pickers/CollectionPicker";
+import { useToast } from "metabase/common/hooks";
+import { getIcon } from "metabase/lib/icon";
+import {
+  FIXED_METABOT_ENTITY_IDS,
+  FIXED_METABOT_IDS,
+} from "metabase/metabot/constants";
+import { PLUGIN_MODERATION } from "metabase/plugins";
+import {
+  Box,
+  Button,
+  Flex,
+  Icon,
+  Loader,
+  Stack,
+  Switch,
+  Text,
+} from "metabase/ui";
+import type {
+  Collection,
+  CollectionEssentials,
+  EnterpriseSettingKey,
+  MetabotInfo,
+} from "metabase-types/api";
 
+import { MetabotPromptSuggestionPane } from "./MetabotAdminSuggestedPrompts";
+import { MetabotNavPane } from "./MetabotNavPane";
 import { MetabotSQLGenerationSettingsSection } from "./MetabotSQLGenerationSettingsSection";
+import { useMetabotIdPath } from "./utils";
 
 export function getAdminRoutes() {
   // NOTE: weird cypress bundle error happens if this `getAdminRoutes` fn
   // is imported into our ai plugin file as the default OSS implementation.
-  return (
-    PLUGIN_METABOT.getAdminRoutes?.() ?? [
-      <IndexRoute key="index" component={MetabotAdminPage} />,
-      <Redirect key="redirect" from="*" to="/admin/metabot" />,
-    ]
-  );
+  return [
+    <IndexRoute key="index" component={MetabotAdminPage} />,
+    <Route
+      key="route"
+      path="sql-generation"
+      component={MetabotAdminSqlGenPage}
+    />,
+    <Route key="route" path=":metabotId" component={MetabotAdminPage} />,
+    <Redirect key="redirect" from="*" to="/admin/metabot" />,
+  ];
 }
 
-function MetabotAdminPage() {
+function MetabotAdminSqlGenPage() {
   return (
     <AdminSettingsLayout sidebar={<MetabotNavPane />}>
       <ErrorBoundary>
@@ -33,17 +72,273 @@ function MetabotAdminPage() {
   );
 }
 
-function MetabotNavPane() {
+export function MetabotAdminPage() {
+  const metabotId = useMetabotIdPath() ?? FIXED_METABOT_IDS.DEFAULT;
+  const { data, isLoading, error } = useListMetabotsQuery();
+
+  const metabot = data?.items?.find((bot) => bot.id === metabotId);
+
+  if (isLoading || !metabot) {
+    return (
+      <LoadingAndErrorWrapper
+        loading={isLoading || !data}
+        error={match({ isLoading, error, metabot })
+          .with(
+            { isLoading: false, error: P.not(null) },
+            () => t`Error fetching Metabots`,
+          )
+          .with({ isLoading: false, metabot: undefined }, () => t`Not found.`)
+          .otherwise(() => null)}
+      />
+    );
+  }
+
+  const isEmbedMetabot =
+    metabot.entity_id === FIXED_METABOT_ENTITY_IDS.EMBEDDED;
+
+  const enabledSettingKey: EnterpriseSettingKey = isEmbedMetabot
+    ? "embedded-metabot-enabled?"
+    : "metabot-enabled?";
+
   return (
-    <Flex direction="column" flex="0 0 auto">
-      <AdminNavWrapper>
-        <AdminNavItem
-          key="sql"
-          icon="sql"
-          label={t`SQL Generation`}
-          path="/admin/metabot"
+    <AdminSettingsLayout sidebar={<MetabotNavPane />}>
+      <ErrorBoundary>
+        <SettingsSection key={metabotId}>
+          <Box>
+            <SettingHeader
+              id="configure-metabot"
+              title={c("{0} is the name of an AI assistant")
+                .t`Configure ${metabot.name}`}
+              description={c("{0} is the name of an AI assistant") // eslint-disable-next-line metabase/no-literal-metabase-strings -- admin ui
+                .t`${metabot.name} is Metabase's AI agent. To help ${metabot.name} more easily find and focus on the data you care about most, configure what content it should be able to access or use to create queries.`}
+            />
+            {isEmbedMetabot && (
+              <Text c="text-secondary" maw="40rem">
+                {t`If you're embedding the Metabot component in an app, you can specify a different collection that embedded Metabot is allowed to use for creating queries.`}
+              </Text>
+            )}
+            <MetabotEnabledToggle
+              settingKey={enabledSettingKey}
+              isEmbedMetabot={isEmbedMetabot}
+            />
+          </Box>
+
+          <MetabotSettingsBody settingKey={enabledSettingKey}>
+            <PLUGIN_MODERATION.MetabotVerifiedContentConfigurationPane
+              metabot={metabot}
+            />
+
+            <MetabotCollectionConfigurationPane
+              metabot={metabot}
+              title={
+                isEmbedMetabot
+                  ? undefined
+                  : t`Collection for natural language querying`
+              }
+            />
+
+            <MetabotPromptSuggestionPane metabot={metabot} />
+          </MetabotSettingsBody>
+        </SettingsSection>
+      </ErrorBoundary>
+    </AdminSettingsLayout>
+  );
+}
+
+function MetabotEnabledToggle({
+  settingKey,
+  isEmbedMetabot = false,
+}: {
+  settingKey: EnterpriseSettingKey;
+  isEmbedMetabot?: boolean;
+}) {
+  const { value, updateSetting, isLoading } = useAdminSetting(settingKey);
+
+  const handleToggle = async (checked: boolean) => {
+    await updateSetting({ key: settingKey, value: checked });
+  };
+
+  return (
+    <Box mt="lg">
+      <SettingHeader
+        id="enable-metabot"
+        title={isEmbedMetabot ? t`Enable Embedded Metabot` : t`Enable Metabot`}
+        description={
+          isEmbedMetabot
+            ? undefined
+            : t`Metabot is Metabase's AI assistant. When enabled, Metabot will be available to help users create queries, analyze data, and answer questions about your data.` // eslint-disable-line metabase/no-literal-metabase-strings -- admin UI
+        }
+      />
+      <Flex align="center" gap="md" mt="md">
+        <Switch
+          data-testid="metabot-enabled-toggle"
+          checked={!!value}
+          onChange={(e) => handleToggle(e.target.checked)}
+          disabled={isLoading}
+          w="auto"
+          size="sm"
         />
-      </AdminNavWrapper>
+        <Text c={value ? "text-primary" : "text-secondary"} fw="500">
+          {isEmbedMetabot
+            ? value
+              ? t`Embedded Metabot is enabled`
+              : t`Embedded Metabot is disabled`
+            : value
+              ? t`Metabot is enabled`
+              : t`Metabot is disabled`}
+        </Text>
+      </Flex>
+    </Box>
+  );
+}
+
+function MetabotSettingsBody({
+  settingKey,
+  children,
+}: {
+  settingKey: EnterpriseSettingKey;
+  children: ReactNode;
+}) {
+  const { value } = useAdminSetting(settingKey);
+  const isDisabled = value === false;
+
+  return (
+    <Stack
+      gap="lg"
+      opacity={isDisabled ? 0.4 : 1}
+      aria-disabled={isDisabled || undefined}
+      style={{
+        pointerEvents: isDisabled ? "none" : "auto",
+        transition: "opacity 150ms ease",
+      }}
+    >
+      {children}
+    </Stack>
+  );
+}
+
+function MetabotCollectionConfigurationPane({
+  metabot,
+  title,
+}: {
+  metabot: MetabotInfo;
+  title?: string;
+}) {
+  const metabotId = metabot.id;
+  const metabotName = metabot.name;
+
+  const {
+    data: collection,
+    isLoading,
+    error,
+  } = useGetCollectionQuery({
+    id: metabot.collection_id ?? "root",
+  });
+
+  const [updateMetabot, { isLoading: isUpdating }] = useUpdateMetabotMutation();
+  const [isOpen, { open, close }] = useDisclosure(false);
+  const [sendToast] = useToast();
+
+  if (isLoading || !collection || error) {
+    return (
+      <LoadingAndErrorWrapper
+        loading={isLoading}
+        error={error ? t`Error fetching Metabot configuration` : null}
+      />
+    );
+  }
+
+  const handleUpdateCollectionId = async (
+    newEntity: Pick<MetabotInfo, "id" | "name">,
+  ) => {
+    close();
+    const result = await updateMetabot({
+      id: metabotId,
+      collection_id: canonicalCollectionId(newEntity.id),
+    });
+
+    if (result.error) {
+      sendToast({
+        message: t`Error setting ${newEntity.name}`,
+        icon: "warning",
+      });
+    }
+  };
+
+  const defaultTitle = c("{0} is the name of an AI assistant")
+    .t`Collection ${metabotName} can use`;
+
+  return (
+    <Box>
+      <SettingHeader id="allow-metabot" title={title ?? defaultTitle} />
+      <CollectionInfo collection={collection} />
+      <Flex gap="md" mt="md">
+        <Button onClick={open} leftSection={isUpdating && <Loader size="xs" />}>
+          {match({ isMutating: isUpdating, collection })
+            .with({ isMutating: true }, () => t`Updating collection...`)
+            .with({ collection: undefined }, () => t`Pick a collection`)
+            .otherwise(() => t`Pick a different collection`)}
+        </Button>
+      </Flex>
+      {isOpen && (
+        <CollectionPickerModal
+          title={t`Select a collection`}
+          value={{
+            id: collection?.id ?? null,
+            model: "collection",
+          }}
+          onChange={(item) =>
+            handleUpdateCollectionId(item as Pick<MetabotInfo, "id" | "name">)
+          }
+          onClose={close}
+          options={{
+            hasRootCollection: true,
+            hasPersonalCollections: false,
+          }}
+        />
+      )}
+    </Box>
+  );
+}
+
+function CollectionInfo({ collection }: { collection: Collection }) {
+  const { data: collectionInfo } = useGetCollectionQuery(
+    collection?.id ? { id: collection.id } : skipToken,
+  );
+
+  if (!collection || !collectionInfo) {
+    return null;
+  }
+
+  const parent = collectionInfo?.effective_ancestors?.slice(-1)?.[0];
+
+  return (
+    <Flex align="center" gap="sm" c="text-tertiary" mb="sm">
+      {parent && (
+        <>
+          <CollectionDisplay collection={parent} />
+          <Text c="text-tertiary" fw="bold">
+            /
+          </Text>
+        </>
+      )}
+      <CollectionDisplay collection={collectionInfo} />
     </Flex>
   );
 }
+
+const CollectionDisplay = ({
+  collection,
+}: {
+  collection: CollectionEssentials;
+}) => {
+  const icon = getIcon({ model: "collection", ...collection });
+  return (
+    <Flex align="center" gap="sm">
+      <Icon {...icon} c={icon.color ?? "brand"} />
+      <Text c="text-secondary" fw="bold">
+        {collection.name}
+      </Text>
+    </Flex>
+  );
+};
