+++ frontend/src/metabase/metabot/components/MetabotAdmin/MetabotAdminSuggestedPrompts.tsx
@@ -13,6 +13,7 @@ import { PaginationControls } from "metabase/common/components/PaginationControl
 import { Table } from "metabase/common/components/Table";
 import { useToast } from "metabase/common/hooks";
 import { usePagination } from "metabase/common/hooks/use-pagination";
+import * as Urls from "metabase/lib/urls";
 import { FIXED_METABOT_IDS } from "metabase/metabot/constants";
 import {
   ActionIcon,
@@ -25,7 +26,6 @@ import {
   Skeleton,
   Tooltip,
 } from "metabase/ui";
-import * as Urls from "metabase-enterprise/urls";
 import type { MetabotInfo, SuggestedMetabotPrompt } from "metabase-types/api";
 
 export const PAGE_SIZE = 10;
