+++ frontend/src/metabase/metabot/components/MetabotAdmin/MetabotAdminSuggestedPrompts.unit.spec.tsx
