+++ frontend/src/metabase/metabot/components/MetabotInlineSQLPrompt/useInlineSQLPrompt.tsx
@@ -15,8 +15,10 @@ import { skipToken, useExtractTablesQuery } from "metabase/api";
 import { useHasTokenFeature } from "metabase/common/hooks";
 import { useDebouncedValue } from "metabase/common/hooks/use-debounced-value";
 import { useRegisterMetabotContextProvider } from "metabase/metabot/context";
-import { useLlmSqlGenerationEnabled } from "metabase/metabot/hooks";
-import { PLUGIN_METABOT } from "metabase/plugins";
+import {
+  useLlmSqlGenerationEnabled,
+  useMetabotSQLSuggestion,
+} from "metabase/metabot/hooks";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
 import type {
@@ -130,7 +132,7 @@ export function useInlineSQLPrompt(
     reset: resetSuggestionState,
     reject,
     suggestionModels,
-  } = PLUGIN_METABOT.useMetabotSQLSuggestion({
+  } = useMetabotSQLSuggestion({
     databaseId,
     bufferId,
     onGenerated,
